+++ R/cbs_extract_table_id.R
@@ -4,7 +4,7 @@
 #' @return `character` with id, will be NA if not found.
 #' @family data retrieval
 cbs_extract_table_id <- function(url, ...){
-  toc <- cbs_get_toc(...)
+  toc <- cbs_get_datasets(...)
   
   if (url %in% toc$Identifier){
     return(url)