+++ example/cbs_search.R
@@ -2,13 +2,20 @@ if (interactive()){
   # search for tables containing the word birth
   
   ds_en <- cbs_search("Birth", language="en")
-  head(ds_en$Title, 3)
+  ds_en[1:3, c("Identifier", "ShortTitle")]
   
   # or in Dutch
   
-  ds_nl <- cbs_search(c("geboorte,sterfte"), language="nl")
-  head(ds_nl$Title, 3)
+  ds_nl <- cbs_search(c("geboorte"), language="nl")
+  ds_nl[1:3, c("Identifier", "ShortTitle")]
   
+  # Search in an other catalog
+  ds_rivm <- cbs_search(c("geboorte"), catalog = "RIVM", language="nl")
+  ds_rivm[1:3, c("Identifier", "ShortTitle")]
+  
+  # search in all catalogs
+  ds_all <- cbs_search(c("geboorte"), catalog = NULL, language="nl")
+
   # docs
   docs <- cbs_search(c("geboorte,sterfte"), language="nl", format="docs")
   names(docs)