+++ NEWS.md
@@ -4,6 +4,11 @@
 
 * removed internal `setwd()` call, fix for issue #28, thx to @jeroenadema 
 
+* added support for other catalogs: 
+
+        - `cbs_get_catalogs` and `cbs_get_datasets`
+        - `cbs_get_data` and `cbs_get_meta` and the like now have a `catalog` argument
+
 # version 0.4.2
 
 * added weeks to cbs_add_date_column, issue #25, thanks to @RichardFromHolland